face-reg-db/requirements.txt
tensorflow==2.2.0
keras==2.3.1
h5py==2.10.0
mtcnn==0.1.0
scikit-learn==0.24.2
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.5.64
psycopg2-binary==2.9.9

